requirements.txt
Django==1.6.5
South==1.0
distribute==0.6.34
djangorestframework==2.3.14
requests==2.3.0
wsgiref==0.1.2
